pyproject.toml
[tool.poetry]
name = "benchmark"
version = "0.1.0"
description = ""
authors = ["Jordan Begg"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.2"
uvicorn = "^0.20.0"
alembic = "^1.9.4"
python-dotenv = "^0.21.1"
sqlmodel = "^0.0.8"
pre-commit = "^3.1.1"
pyodbc = "^5.2.0"
fastapi = "^0.115.5"
pyjwt = "^2.10.0"
passlib = "^1.7.4"
bcrypt = "4.0.1"
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pytest = "^7.2.2"
requests = "^2.28.2"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
